+++ packages/vue/test/base/src/views/tabs-similar-prefixes/Home3.vue
@@ -1,0 +1,16 @@
+<template>
+  <ion-page data-pageid="home3">
+    <ion-header>
+      <ion-toolbar>
+        <ion-title>Home 3</ion-title>
+      </ion-toolbar>
+    </ion-header>
+    <ion-content>
+      <div data-testid="home3-content">Home 3 Content</div>
+    </ion-content>
+  </ion-page>
+</template>
+
+<script setup lang="ts">
+import { IonContent, IonHeader, IonPage, IonTitle, IonToolbar } from '@ionic/vue';
+</script>